# api_gateway/docs/swagger.yaml
basePath: /
definitions:
...
  dto.Response:
    properties:
      data:
        x-order: "1"
      message:
        type: string
        x-order: "0"
    type: object
...
  dto.SwaggerResponseRegister:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/models.User'
        x-order: "1"
      message:
        type: string
        x-order: "0"
    type: object
...
  dto.UserLogin:
    properties:
      email:
        type: string
        x-order: "0"
      password:
        type: string
        x-order: "1"
    required:
    - email
    - password
    type: object
  dto.UserRegister:
    properties:
      email:
        type: string
        x-order: "2"
      first_name:
        type: string
        x-order: "0"
      last_name:
        type: string
        x-order: "1"
      password:
        type: string
        x-order: "3"
    required:
    - email
    - first_name
    - password
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
...
  utils.ErrResponse:
    properties:
      detail:
        type: string
        x-order: "2"
      status:
        type: integer
        x-order: "0"
      type:
        type: string
        x-order: "1"
    type: object
...
host: 3ee9-103-18-34-211.ngrok-free.app
info:
  contact:
    email: muhlisiqbalutomoo@gmail.com
    name: The developer
  description: Online motorcycle washing service built with microservices that integrates
    user, washer, and admin.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Fox Wash
  version: 0.0.1
paths:
  /admins/washer-activation/{email}:
    post:
      consumes:
      - application/json
      description: Activation washer by admin for hired washer as team
      parameters:
      - description: Email Washer
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Washer Activation
      tags:
      - admin
  /users/login:
    post:
      consumes:
      - application/json
      description: Login users and embeds a JWt-Auth in cookie
      parameters:
      - description: User login details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Log in
      tags:
      - all user
...
  /users/register/admin:
    post:
      consumes:
      - application/json
      description: Register a new user with the role 'admin'
      parameters:
      - description: Admin registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SwaggerResponseRegister'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Register a new admin
      tags:
      - admin
  /users/register/user:
    post:
      consumes:
      - application/json
      description: Register a new user with the role 'user'
      parameters:
      - description: User registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SwaggerResponseRegister'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Register a new user
      tags:
      - customer
  /users/register/washer:
    post:
      consumes:
      - application/json
      description: Register a new user with the role 'washer'
      parameters:
      - description: Washer registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SwaggerResponseRegister'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Register a new washer
      tags:
      - washer
  /users/verify/{userid}/{token}:
    get:
      consumes:
      - application/json
      description: Verify the user registration using unique token sent to the registered
        email
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        type: integer
      - description: Verification token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Verify user credentials
      tags:
      - all user
...
swagger: "2.0"